fix: extend keepdims fix to all reduction operations
Extended the keepdims fix to all remaining reduction operations:
- ReduceAMax (np.amax, np.max)
- ReduceAMin (np.amin, np.min)
- ReduceProduct (np.prod)
- ReduceStd (np.std)
- ReduceVar (np.var)

Also fixed np.amax/np.amin API layer which ignored keepdims when axis=null.

Added comprehensive parameterized test covering all reductions with
multiple dtypes (Int32, Int64, Single, Double, Int16, Byte) to prevent
regression.

All 7 reduction functions now correctly preserve dimensions with
keepdims=true, matching NumPy 2.x behavior.

+++ src/NumSharp.Core/Backends/Default/Math/Reduction/Default.Reduction.Product.cs
@@ -18,7 +18,7 @@ namespace NumSharp.Backends
                 if (axis_ == null)
                 {
                     var r = NDArray.Scalar((typeCode ?? arr.typecode).GetOneValue());
-                    if (keepdims) { var ks = new int[arr.ndim]; for (int i = 0; i < arr.ndim; i++) ks[i] = 1; r.Storage.Reshape(new Shape(ks)); }
+                    if (keepdims) { var ks = new long[arr.ndim]; for (int i = 0; i < arr.ndim; i++) ks[i] = 1; r.Storage.Reshape(new Shape(ks)); }
                     return r;
                 }
                 var axis = NormalizeAxis(axis_.Value, arr.ndim);
@@ -26,7 +26,7 @@ namespace NumSharp.Backends
                 var result = np.ones(new Shape(resultShape), typeCode ?? arr.GetTypeCode.GetAccumulatingType());
                 if (keepdims)
                 {
-                    var ks = new int[arr.ndim];
+                    var ks = new long[arr.ndim];
                     for (int d = 0, sd = 0; d < arr.ndim; d++) ks[d] = (d == axis) ? 1 : resultShape[sd++];
                     result.Storage.Reshape(new Shape(ks));
                 }
@@ -40,7 +40,7 @@ namespace NumSharp.Backends
             {
                 var result = prod_elementwise_il(arr, typeCode);
                 var r = NDArray.Scalar(result);
-                if (keepdims) { var ks = new int[arr.ndim]; for (int i = 0; i < arr.ndim; i++) ks[i] = 1; r.Storage.Reshape(new Shape(ks)); }
+                if (keepdims) { var ks = new long[arr.ndim]; for (int i = 0; i < arr.ndim; i++) ks[i] = 1; r.Storage.Reshape(new Shape(ks)); }
                 else if (!r.Shape.IsScalar && r.Shape.size == 1 && r.ndim == 1) r.Storage.Reshape(Shape.Scalar);
                 return r;
             }
